+++ src/main/java/beforespring/socialfeed/member/service/MemberService.java
@@ -1,7 +1,7 @@
 package beforespring.socialfeed.member.service;
 
-import static beforespring.socialfeed.member.controller.dto.ConfirmTokenDto.ConfirmTokenRequest;
-import static beforespring.socialfeed.member.controller.dto.CreateMemberDto.CreateMemberRequest;
+import static beforespring.socialfeed.member.service.dto.ConfirmTokenDto.ConfirmTokenRequest;
+import static beforespring.socialfeed.member.service.dto.CreateMemberDto.CreateMemberRequest;
 
 import beforespring.socialfeed.jwt.domain.AuthToken;
 import beforespring.socialfeed.member.service.dto.PasswordAuth;
