+++ src/main/java/beforespring/socialfeed/web/api/v1/content/ContentControllerAdvice.java
@@ -1,6 +1,6 @@
-package beforespring.socialfeed.content.controller;
+package beforespring.socialfeed.web.api.v1.content;
 
-import beforespring.socialfeed.content.controller.response.ErrorResponse;
+import beforespring.socialfeed.web.api.v1.response.ErrorResponse;
 import beforespring.socialfeed.content.service.exception.ContentNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
